feat: confirm folder deletion

+++ src/ui/files/Files.jsx
@@ -61,6 +61,7 @@ class Files extends React.Component {
     }
 
     @action deleteFolder = folder => {
+        if (!confirm(t('title_deleteFolder'))) return;
         this.currentFolder = folder.parent;
         fileStore.fileFolders.deleteFolder(folder);
         fileStore.fileFolders.save();
